Move custom shaders to shaders subdirectory.
