+++ app.js
@@ -16,12 +16,12 @@ import {Axes} from "./axes.js";
 import {Rect} from "./rect.js";
 import {Colors} from "./colors.js";
 import {Coords} from "./coords.js";
-import {DepthShader} from "./depth_shader.js";
+import {DepthShader} from "./shaders/depth_shader.js";
 import {EventTracker} from "./event_tracker.js";
 import {Extruder} from "./extruder.js";
 import {FetchQueue} from "./fetch_queue.js";
 import {Ground} from "./ground.js";
-import {HazeShader} from "./haze_shader.js";
+import {HazeShader} from "./shaders/haze_shader.js";
 import {Settings} from "./settings.js";
 import {SkyBox} from "./skybox.js";
 import {Util} from "./util.js";
